minor code refactoring

+++ demos/set_sample.c
@@ -1,9 +1,10 @@
 /**
- * Copyright (c) 2020 Surgams
- *
- * This library is free software; you can redistribute it and/or modify it
- * under the terms of the BSD license. See LICENSE for details.
- **/
+*  Copyright (c) 2020 Surgams
+*
+*  This library is free software; you can redistribute it and/or modify it
+*  under the terms of the BSD license. See LICENSE for details.
+*
+**/
 
 #include <string.h>
 #include "../src/ds_set.h"
@@ -48,7 +49,7 @@ int main() {
 
     Data *data, *data2;
 
-    ushort ret_code =-1;
+    uint8_t ret_code =-1;
 
     /*Initialise data*/
     if ((data = (Data *) malloc(sizeof(Data))) == NULL)
