+++ src/ds_hash.h
@@ -1,13 +1,15 @@
 /**
- * Copyright (c) 2020 Surgams
- *
- * This library is free software; you can redistribute it and/or modify it
- * under the terms of the BSD license. See LICENSE for details.
- **/
+*  Copyright (c) 2020 Surgams
+*
+*  This library is free software; you can redistribute it and/or modify it
+*  under the terms of the BSD license. See LICENSE for details.
+*
+**/
 
 #ifndef DS_HASH_H
 #define DS_HASH_H
 
+#include <stdint.h>
 #include "ds_list.h"
 
 /* Define structure for the hash node*/
@@ -34,12 +36,12 @@ typedef DS_Hash_Table * DS_Hash_Table_Ptr;
 
 /* Public Interface */
 
-ushort ds_hash_init (DS_Hash_Table_Ptr *htbl, int buckets, Compare compare, Destroy destroy);
+uint8_t ds_hash_init (DS_Hash_Table_Ptr *htbl, int buckets, Compare compare, Destroy destroy);
 
-ushort ds_hash_free (DS_Hash_Table_Ptr *htbl);
-ushort ds_hash_put (DS_Hash_Table_Ptr htbl, const char *key, const void *data);
-ushort ds_hash_remove (DS_Hash_Table_Ptr htbl, const char *key, void **data);
-ushort ds_hash_get (const DS_Hash_Table_Ptr htbl, const char *key, void **data);
+uint8_t ds_hash_free (DS_Hash_Table_Ptr *htbl);
+uint8_t ds_hash_put (DS_Hash_Table_Ptr htbl, const char *key, const void *data);
+uint8_t ds_hash_remove (DS_Hash_Table_Ptr htbl, const char *key, void **data);
+uint8_t ds_hash_get (const DS_Hash_Table_Ptr htbl, const char *key, void **data);
 
 #define ds_hash_size(htbl) ((htbl)->size)
