+++ src/ds_stack.h
@@ -1,14 +1,15 @@
 /**
- * Copyright (c) 2020 Surgams
- *
- * This library is free software; you can redistribute it and/or modify it
- * under the terms of the BSD license. See LICENSE for details.
- **/
+*  Copyright (c) 2020 Surgams
+*
+*  This library is free software; you can redistribute it and/or modify it
+*  under the terms of the BSD license. See LICENSE for details.
+*
+**/
 
 #ifndef DS_STACK_H
 #define DS_STACK_H
 
-
+#include <stdint.h>
 #include "ds_list.h"
 
 /* Implement stacks as linked lists */
@@ -21,8 +22,8 @@ typedef DS_List_Ptr DS_Stack_Ptr;
 #define ds_stack_purge ds_list_purge
 #define ds_stack_free ds_list_free
 
-ushort ds_stack_push(DS_Stack_Ptr, const void *);
-ushort ds_stack_pop(DS_Stack_Ptr, void **);
+uint8_t ds_stack_push(DS_Stack_Ptr, const void *);
+uint8_t ds_stack_pop(DS_Stack_Ptr, void **);
 
 #define ds_stack_peek(stack)((stack)->head == NULL ? NULL : (stack)->head->data)
 #define ds_stack_size ds_list_size
